creating product details component
