--- client/src/components/ProductImages/index.js
@@ -1,12 +1,0 @@
-import React from "react";
-import ProductList from "../ProductList";
-
-function Images() {
-    return(
-        <div>
-
-        </div>
-    )
-}
-
-export default Images;
