--- client/src/components/ProductPage/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function ProductPage() {
-    return(
-       <div>
-
-       </div>
-    )
-}
-
-export default ProductPage;
